requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
python-dotenv==1.0.1
asyncpg>=0.30.0
websockets==13.1
azure-cognitiveservices-speech==1.40.0
silero-vad==5.1.2
httpx==0.27.2
pydantic==2.9.2
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.10
google-cloud-vision==3.8.0
Pillow==11.0.0
pdf2image==1.17.0
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.24.0